Renamed tests and created pkg and co_coders inside internal
