+++ internal/co_coders/user/user.go
@@ -1,6 +1,6 @@
 package user
 
-import "github.com/clean-code-projects/co-coders-api/internal/criteria"
+import "github.com/clean-code-projects/co-coders-api/internal/co_coders/criteria"
 
 // User ..
 type User struct {
